Restructured them the notification related files, and removed the console.log
